delivery of DOD GOSS OWF v7.17.2.0-RC1
