--- application.properties
@@ -1,4 +1,0 @@
-#Grails Metadata file
-#Wed Jan 08 14:20:21 EST 2014
-app.grails.version=2.3.7
-app.name=ozone-messaging
