+++ build.gradle
@@ -1,0 +1,88 @@
+buildscript {
+    ext {
+        grailsVersion = '3.3.2'
+        gormVersion = '6.1.8.RELEASE'
+    }
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        maven { url "https://plugins.gradle.org/m2/" }
+        maven { url 'https://repo.grails.org/grails/core' }
+        jcenter()
+    }
+    dependencies {
+        classpath group: 'io.spring.gradle', name: 'dependency-management-plugin', version: '1.0.4.RELEASE'
+
+        classpath group: 'org.grails', name: 'grails-gradle-plugin', version: grailsVersion
+    }
+}
+
+
+group 'org.ozoneplatform'
+version '1.19.1-0-RC1'
+
+
+apply plugin: 'io.spring.dependency-management'
+
+apply plugin: 'idea'
+apply plugin: 'maven'
+apply plugin: 'java'
+
+apply plugin: 'org.grails.grails-plugin'
+apply plugin: 'org.grails.grails-plugin-publish'
+
+
+bootRun {
+    jvmArgs('-Dspring.output.ansi.enabled=always')
+    addResources = true
+}
+
+// Skip integration tests during build/assemble phase
+check.dependsOn -= integrationTest
+
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    maven { url 'https://repo.grails.org/grails/core' }
+    maven { url 'http://repository.springsource.com/maven/bundles/release' }
+    maven { url 'http://repository.springsource.com/maven/bundles/external' }
+    maven { url 'https://packages.atlassian.com/maven-3rdparty-legacy' }
+}
+
+dependencyManagement {
+    imports {
+        mavenBom 'org.grails:grails-bom:' + grailsVersion
+        mavenBom 'org.ozoneplatform:ozone-classic-bom:7.17.2-0-SNAPSHOT'
+    }
+
+    applyMavenExclusions false
+}
+
+dependencies {
+    // Spring Boot
+    provided group: 'org.springframework.boot', name: 'spring-boot-autoconfigure'
+    provided group: 'org.springframework.boot', name: 'spring-boot-starter-logging'
+
+    // Grails
+    compile group: 'org.grails', name: 'grails-core'
+
+    console group: 'org.grails', name: 'grails-console'
+
+    profile group: 'org.grails.profiles', name: 'plugin'
+
+    provided group: 'org.grails', name: 'grails-plugin-services'
+    provided group: 'org.grails', name: 'grails-plugin-domain-class'
+
+    testCompile group: 'org.grails', name: 'grails-plugin-testing'
+    testCompile group: 'org.grails', name: 'grails-gorm-testing-support'
+
+    // Other
+    compile group: 'org.igniterealtime.smack', name: 'smack'
+    compile group: 'org.igniterealtime.smack', name: 'smackx'
+}
+
+
+task wrapper(type: Wrapper) {
+    gradleVersion = '4.2.1'
+}
