+++ gradle.properties
@@ -1,0 +1,5 @@
+# Enable Gradle daemon
+org.gradle.daemon=true
+
+# Enable parallel build (experimental)
+org.gradle.parallel=true
