--- grails-app/conf/BuildConfig.groovy
@@ -1,35 +1,0 @@
-grails.project.class.dir = "target/classes"
-grails.project.test.class.dir = "target/test-classes"
-grails.project.test.reports.dir = "target/test-reports"
-grails.project.dependency.resolver = "maven"
-
-grails.project.dependency.resolution = {
-    // inherit Grails' default dependencies
-    inherits("global") {
-        // uncomment to disable ehcache
-        // excludes 'ehcache'
-    }
-    log "warn" // log level of Ivy resolver, either 'error', 'warn', 'info', 'debug' or 'verbose'
-    repositories {
-        grailsPlugins()
-        grailsHome()
-        mavenLocal()
-        mavenRepo 'http://repository.opencastproject.org/nexus/content/repositories/public'
-        grailsCentral()
-        mavenCentral()
-    }
-    dependencies {
-        // specify dependencies here under either 'build', 'compile', 'runtime', 'test' or 'provided' scopes eg.
-
-        // runtime 'mysql:mysql-connector-java:5.1.13'
-        compile( 'org.igniterealtime.smack:smack:3.3.1')
-        compile( 'org.igniterealtime.smack:smackx:3.3.1')
-
-    }
-
-    plugins{
-        compile (':release:3.0.1'){
-            export=false
-        }
-    }
-}
