--- grails-app/conf/UrlMappings.groovy
@@ -1,13 +1,0 @@
-class UrlMappings {
-
-	static mappings = {
-		"/$controller/$action?/$id?"{
-			constraints {
-				// apply constraints here
-			}
-		}
-
-		"/"(view:"/index")
-		"500"(view:'/error')
-	}
-}
