+++ grails-app/conf/application.yml
@@ -1,0 +1,39 @@
+grails:
+    profile: plugin
+    codegen:
+        defaultPackage: org.ozoneplatform.messaging
+    spring:
+        transactionManagement:
+            proxies: false
+    gorm:
+        reactor:
+            # Whether to translate GORM events into Reactor events
+            # Disabled by default for performance reasons
+            events: false
+info:
+    app:
+        name: '@info.app.name@'
+        version: '@info.app.version@'
+        grailsVersion: '@info.app.grailsVersion@'
+spring:
+    main:
+        banner-mode: "off"
+    groovy:
+        template:
+            check-template-location: false
+
+# Spring Actuator Endpoints are Disabled by Default
+endpoints:
+    enabled: false
+    jmx:
+        enabled: true
+
+environments:
+    test:
+        notifications:
+            xmpp:
+                username: ''
+                password: ''
+                host: ''
+                room: ''
+                port: 5222
