+++ grails-app/init/org/ozoneplatform/messaging/Application.groovy
@@ -1,0 +1,12 @@
+package org.ozoneplatform.messaging
+
+import grails.boot.*
+import grails.boot.config.GrailsAutoConfiguration
+import grails.plugins.metadata.*
+
+@PluginSource
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
