+++ grails-app/init/org/ozoneplatform/messaging/BootStrap.groovy
@@ -1,0 +1,9 @@
+package org.ozoneplatform.messaging
+
+class BootStrap {
+
+    def init = { servletContext ->
+    }
+    def destroy = {
+    }
+}
