--- pom.xml
@@ -1,55 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<project xmlns='http://maven.apache.org/POM/4.0.0' xmlns:xsi='http://www.w3.org/2001/XMLSchema-instance' xsi:schemaLocation='http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd'>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.ozoneplatform</groupId>
-  <artifactId>ozone-messaging</artifactId>
-  <packaging>jar</packaging>
-  <version>1.19</version>
-  <name>OZONE Messaging plugin</name>
-  <description>\
-        Common messaging components.
-    </description>
-
-  <url>https://github.com/ozoneplatform/owf-messaging</url>
-
-  <developers>
-    <developer>
-      <name>OZONE Team</name>
-      <email>goss-support@owfgoss.org</email>
-    </developer>
-  </developers>
-
-  <licenses>
-        <license>
-            <name>Apache License, Version 2.0</name>
-            <url>http://www.apache.org/licenses/LICENSE-2.0</url>
-        </license>
-    </licenses>
-    <scm>
-        <connection>scm:git:git@github.com:ozoneplatform/owf-messaging.git</connection>
-        <developerConnection>scm:git:git@github.com:ozoneplatform/owf-messaging.git</developerConnection>
-        <url>git@github.com:ozoneplatform/owf-messaging.git</url>
-    </scm>
-  <dependencies>
-    <dependency>
-      <groupId>org.igniterealtime.smack</groupId>
-      <artifactId>smack</artifactId>
-      <version>3.3.1</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.igniterealtime.smack</groupId>
-      <artifactId>smackx</artifactId>
-      <version>3.3.1</version>
-      <scope>compile</scope>
-    </dependency>
-  </dependencies>
-  <repositories>
-    <repository>
-      <id>nexus.content.repositories.public</id>
-      <name>Maven Nexus Server</name>
-      <url>http://repository.opencastproject.org/nexus/content/repositories/public</url>
-    </repository>
-  </repositories>
-
-</project>
