+++ src/main/groovy/org/ozoneplatform/messaging/connection/XmppConnectionInformation.groovy
