+++ src/main/groovy/org/ozoneplatform/messaging/payload/OzoneMessage.groovy
