--- test/unit/org/ozoneplatform/messaging/service/impl/XmppConfigurationServiceTests.groovy
@@ -1,17 +1,0 @@
-package org.ozoneplatform.messaging.service.impl
-
-import grails.test.*
-
-class XmppConfigurationServiceTests extends GrailsUnitTestCase {
-    protected void setUp() {
-        super.setUp()
-    }
-
-    protected void tearDown() {
-        super.tearDown()
-    }
-
-    void testSomething() {
-
-    }
-}
