--- web-app/WEB-INF/sitemesh.xml
@@ -1,14 +1,0 @@
-<sitemesh>
-    <page-parsers>
-        <parser content-type="text/html"
-            class="org.codehaus.groovy.grails.web.sitemesh.GrailsHTMLPageParser" />
-        <parser content-type="text/html;charset=ISO-8859-1"
-            class="org.codehaus.groovy.grails.web.sitemesh.GrailsHTMLPageParser" />
-        <parser content-type="text/html;charset=UTF-8"
-            class="org.codehaus.groovy.grails.web.sitemesh.GrailsHTMLPageParser" />            
-    </page-parsers>
-
-    <decorator-mappers>
-        <mapper class="org.codehaus.groovy.grails.web.sitemesh.GrailsLayoutDecoratorMapper" />
-    </decorator-mappers>
-</sitemesh>
